requirements.txt
openpyxl==3.0.7
pyqt==5.15.4
pymysql==1.0.2